Dynamax smoothers
